CESI::_CheckUtf16Eol の第2引数の型を int から size_t に変更 x64 ビルドで警告が出る数を減らす為の変更 関数の呼び出し元では pr_end-pr という記述でバッファ長を求めており、 pr_end の値の作り方から考えて演算結果が負の値になる事は無いと思われる

+++ sakura_core/charset/CESI.h
@@ -149,9 +149,9 @@ protected:
 	void GetEncodingInfo_latin1( const char* pS, const int nLen );
 	void GetEncodingInfo_meta( const char* pS, const int nLen );
 
-	bool _CheckUtf16Eol( const char* pS, const int nLen, const bool bbig_endian );
-	inline bool _CheckUtf16EolLE( const char* p, const int n ){ return _CheckUtf16Eol( p, n, false ); }
-	inline bool _CheckUtf16EolBE( const char* p, const int n ){ return _CheckUtf16Eol( p, n, true ); }
+	bool _CheckUtf16Eol( const char* pS, const size_t nLen, const bool bbig_endian );
+	inline bool _CheckUtf16EolLE( const char* p, const size_t n ){ return _CheckUtf16Eol( p, n, false ); }
+	inline bool _CheckUtf16EolBE( const char* p, const size_t n ){ return _CheckUtf16Eol( p, n, true ); }
 
 public:
 	//
